Create unit tests for Card component
The feature of decreasing or increasing the amount of products was
tested to verify whether it worked properly or not. The functions
to achieve these features were changed a bit in order to correct the
first mistakes, in which the click events happened so fast that the
event handlers were using obsolete data.

On the other hand, in order to prevent the Card component from
decreasing the amount of products below 0, this condition is also
evaluated within the useEffect hook, as the conditional inside the
decrement function was not enough when the events happen too fast.