Adjust selecting square and create relative coords
By default, the square was not centered with the cursor after clicking,
which was solved by dividing the width and height of the square by two
and substracting those values from the coordinates provided by the
event. Thus, this result was assigned to the top and left rules of the
DIV to center it.

On the other hand, by resizing the window the coordinates were the
character lies changed, leading to impossible conditions to play. This
was addressed by dividing the coordinates provided by the event by
the height and width, according if it was the x or y coordinate. The
result was multiplied by 100 and rounded, leading to consistent
coordinates despite the size of the screen.