Set categoryPreviousIndex state and setCategoryBackground function
The categoryPreviousIndex state stores the previous index used to
add and remove classes from the respective element in the
categoriesArray ref.

The setCategoryBackground function adds and removes classes from the
previous and current selected categories to set the background color.
This function leverages the categoryPreviousIndex to identify what
operation to perform on each category div.