Fix value container width and add media query
The media query changes the prices-range container flex direction to
column and reduces its padding. Additionally, the 'value' container
margin left is set to 0 whereas the margin top is increased to separate
it from the slider. Finally, the slider width was set to 100% and its
size was reduced along with that of the thumb.