Use pseudo-element to add star icon and color on hover
The star was added using the content css property and adding fontAwesome
as the font-family. In addition, in order to change the stars color the
general sibling combinator (~) was employed to select all input siblings
of the current hovered input. On the other hand, in the main container
the flex-direction was set to row-reversed so the stars color would
begin from left to right.