Style review buttons and fine tune card values
The implicit rows created within the card grid was set to prevent images
from overflowing. On the other hand, containers had a hardcoded height
which was removed to allow the html natural responsiveness to kick in.
In addition, the product name was given font-weight and the review
button was given style. As a consequence of the addition of the review
button, images height was also increased.