Update StarRate to only show rate or allow to set new rating
The StarRate component should only display the rating of items in the
Card component. However, when reviewing an item the StarRate component
should allow users to set their score in order send information to
upcoming server. This is done by changing input type to radio when
updating score and checkbox when displaying rating. On the other hand,
a className is added to handle animation according to current context.