Add custom classes to style svg's used in CardIncentive component
The custom classes address the width, height and margin top of svg's in
CardIncentive components since they are stored in an object in the
constants file; therefore, by means of classes each svg is addressed
without having to style each one individually. Likewise, the path
is given a fill color.