Set rating prop and name attribute in object
The rating was given the default value of 0 to indicate users that
a determined item has no reviews in case this prop is not passed.
On the other hand, this prop will contain the overall reviews from users
as well as the rating given by users depending on the component the
StarRate renders.

In addition, the input name attribute is set to 'rating' or the
product's name. The former will be used when creating reviews so that
the name remains constant, and the latter will be used when displaying
the item's review.