Set AuthButtons component, main container height/padding and user object
The AuthButtons component is wrapped in a div container to hide this
section in medium screen devices. Once the breakpoint is exceeded, the
div container shows the authentication buttons.

On the other hand, the user object is imported from the App component
using the useContext hook. This variable will ultimately decide whether
the AuthButtons component should render in the first place.