Add logic to check if passwords match
The repeatPwd and pwdMatch states are created. The former will hold the
input's value regarding the newly created repeat password section.
In addition, a useEffect is used to check if both passwords match and
to store that validation in pwdMatch.

The pwdMatch state is employed to change the repeat password input
border and outline color to convey that passwords do not match.
Likewise, this state prevents the handleSubmit function from calling the
API until the condition is met.