Send email error msg, add numerical validation to pwd and add username
When the email has already been used an object with the msg key is sent
to the client so it can display a message about the situation to the
user. On the other hand, the password is enforced by requesting at least
1 upper letter and one numerical value. Finally, the username field
is added in the creation of the user.