Set up middlewares to perform login process
The login middleware validates the email entered by users; thus,
allowing to call passport authenticate if the email is valid.

The passport authenticate will redirect to other API's depending on the
status of the login process. If the process was successful, then the
check_user_status middleware will be summoned to return the user object,
conversely, the login_failure middleware is summoned to indicate the
client that something went wrong.

Finally, the logout middleware is set to allow users abandoning their
session.