Set cors only in production and add Catch-all routes
Cors should always run when the app is not in production since both
the client and server will be hosted in the same domain. Likewise,
cors is only accepting requests from the client in localhost.

Finally, the catch-all approach is used to serve the html file located
in the public folder to prevent errors coming from refreshing pages, as
React Router is handling client side router.