Remove input's name, rename pwd state and manage API response
The pwd state was rename with 'password' so both the backend and
frontend would be handling the same variable name. On the other hand,
depending on the response from the API, users should see validation
errors if any. Moreover, when the login process is successful, then
the 'user' state is updated and users are navigated to the home page.