Set states and context variables to create Reviews
The states created are 'comment' and 'rating'. The rating updater is
passed as a prop to the StarRate component. On the other hand, the
comment updater is used every time the textarea element changes.

Finally, the API and user variables are set from CartContext to define
the upcoming function to create Reviews.