Tune attributes according to isCustomizable state
Aside from tuning attributes, a single return is used instead of 2.

When the component is customizable then in the checked attribute the
localRating state must be used since the rating prop is undefined. This
will allow to check radio buttons when reviewing products. Moreover,
the 'radio' type is used instead of also using the checkbox type.

On the other hand, when udpdating the rating prop and localRating state
the value must be converted into an int, otherwise the value is a
string, which will lead to problems since this value is compared to
integers using the strict equality operator (===) preventing from
checking radio buttons.