Set logic to update likes in ReviewCard
The way it works is by setting the localLikes state in order to store
the likes prop. This way, the state will be used to render dinamically
the likes count according to the click and whether the user has already
liked that review.

Likewise, a copy of the localLikes state is created within the
updateReviewLikes function to handle the changes and send the current
value, as well as the current data of the review, to the respective API
endpoint so the database is also updated.