Set data attribute to svg's container and add pointer-events-none
The data attribute will be used to distinguish the like from the dislike
svg to handle the respective operation. Thus, the pointer-events-none is
added to each svg in order to prevent these elements from triggering
the event, which would prevent the logic from retrieving the data
attribute set in the respective parent.

On the other hand, the pointer-events-none would disable the
cursor-pointer style in each svg, so it was removed from the svg's and
set in the parent.