Improve likes and dislikes update logic
Firstly, the localLikes and localDislikes states are being handled by
means of sets in the updateLikesDislikes function in order to
search time. Secondly, the removeUserIndex function is
removed since it is no longer required thanks to the set's methods.
Finally, instead of wrapping the whole updateLikesDislikes function
logic with an if to see if user is logged, this if automatically returns
if no user is logged in.