Add dislike functionality and change local likes and dislikes type
The dislike functionality was added by simply calling the
updateReviewLikes function when the dislike svg container is clicked.
Likewise, when the user has disliked a product, then the svg is colored
in orange.

On the other hand, the localLikes and localDislikes states are converted
into sets as soon as the component is rendered to leverage the set's
methods to find, remove and add users to the likes and dislikes fields.