Store reviews in a single state in Shop component
Instead of retrieving specific reviews from the server for each product,
all reviews are stored and passed as props to the Card component. Thus,
the app's performance is improved by not overloading the server with
requests.

On the other hand, in the Card component only reviews regarding the
current product are retrieved using the filter() built-in function to
compute the average of the rating.