Calculate rating for each product in Shop component
The updatedProducts state is set to store the general product structure
with the rating already computed. This state ensures that data remains
in the client, so every calculation now can be made synchronously.

The rating average is calculated with the getAverageRating function that
is summoned as soon as the backend has provided all reviews, which will
update the updatedProducts state that in turn will be filtered so
products matching price and category will be displayed using the
itemsToShow state.