Optimize getAverageRating function
Previously this function iterated the whole reviews array for each
product in the store, which was inefficient in terms of time.

The function now iterates once the reviews array to store each review
with the occurrence count and the sum of the rating according to the id.
Later, each piece of data is recovered by each product in the store
to update the rating.