Set type attribute in StarRate inputs
The type attribute switches between radio and checkbox. The radio type
is used only when users are writing a review; conversely, the checkbox
type is used to display item's rating without allowing users to alter
it.

The reason behind this was that the radio type was not rendering
some inputs to display item's rating, but the checkbox type proved to be
better to keem consistency in the rating of each product when refreshing
the page.