Filter respective product reviews in ReviewForm
The ReviewForm checks if the user has already reviewed the product
by checking that the user id is not already in the reviews of that
specific product. This is done by filtering the global reviews to filter
the product reviews to check for that condition. Likewise, the global
reviews are updated when a review has been created with the help of the
setUpdateReviews, that switches its values to trigger the useEffect in
App.