Update user's cart using handleProductAddition function
This function calls the respective API to update the current user's
cart. Moreover, the function updates the user from the client by
checking whether the product added is already in the cart or is a new
addition. Finally, the outcome is sent to the backend to update
the user data.