Set toggleNavbar function to close navbar
The toggleNavbar function checks that the setToggleNavbar prop is not
undefined prior to using it when clicking any button. The reason of the
toggleNavbar function is that these buttons are also placed in the
Header for large devices; thus, when located these buttons there then
they should not command the navbar state.