Remove console log and add product id in Card key
The product id is now used in the key of the Card component when it
comes to rendering that component with the map function. Previously the
iterator provided by the map function was used to assign the keys,
however this led to unexpected bugs such as rendering the number
of reviews in the wrong products when the filter changed.