Add logic to edit reviews in ReviewForm component
With the help of the 'edit' parameter it is defined when the component
should allow editing or creating a review. For this, the updateReview
function is defined, which is in charge of calling the API endpoint to
modify the rating and the user's comment. Also, the function
handleReviewSubmit is defined to call either the createReview or
updateReview function depending on the context when the component form
is submitted.