Add reviews field in data structure of products
The data structure containing all product related data now has a reviews
field to store the reviews for each item. This is done to preserve the
reviews when refreshing the page in the ProductData component, as
previously the globalReviews state was used to store all the reviews,
but this data was lost when refreshes were performed.