Set getUserStatus function
This function will replace the useEffect to call the check_user_status
API endpoint to check whether the user has an active session. The
getUserStatus uses the newly created setCart function to retrieve
the item's data stored in the client so the application Cart renders
accordingly with the user's data.